+++ components/openapi/src/test/scala/pl/touk/nussknacker/openapi/BaseOpenAPITest.scala
@@ -4,12 +4,12 @@ import cats.data.Validated
 import cats.data.Validated.{Invalid, Valid}
 import org.apache.commons.io.IOUtils
 import pl.touk.nussknacker.engine.api.process.ComponentUseCase
-import pl.touk.nussknacker.engine.api.{Context, ContextId, JobData, MetaData, ProcessVersion, StreamMetaData}
+import pl.touk.nussknacker.engine.api._
 import pl.touk.nussknacker.engine.util.runtimecontext.TestEngineRuntimeContext
 import pl.touk.nussknacker.engine.util.service.EagerServiceWithStaticParametersAndReturnType
 import pl.touk.nussknacker.openapi.enrichers.{SwaggerEnricherCreator, SwaggerEnrichers}
 import pl.touk.nussknacker.openapi.parser.{ServiceParseError, SwaggerParser}
-import sttp.client3.testing.SttpBackendStub
+import sttp.client3.SttpBackend
 
 import java.net.URL
 import java.nio.charset.StandardCharsets
@@ -47,7 +47,7 @@ trait BaseOpenAPITest {
 
   protected def parseToEnrichers(
       resource: String,
-      backend: SttpBackendStub[Future, Any],
+      backend: SttpBackend[Future, Any],
       config: OpenAPIServicesConfig = baseConfig
   ): Map[ServiceName, EagerServiceWithStaticParametersAndReturnType] = {
     val services = parseServicesFromResourceUnsafe(resource, config)