+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/server/RouteProvider.scala
@@ -2,9 +2,10 @@ package pl.touk.nussknacker.ui.server
 
 import akka.http.scaladsl.server.Route
 import cats.effect.{IO, Resource}
-import pl.touk.nussknacker.engine.ConfigWithUnresolvedVersion
+import pl.touk.nussknacker.ui.config.DesignerConfig
 
 trait RouteProvider[R <: Route] {
 
-  def createRoute(config: ConfigWithUnresolvedVersion): Resource[IO, R]
+  def createRoute(config: DesignerConfig): Resource[IO, R]
+
 }