+++ engine/flink/components/base-unbounded/src/main/java/pl/touk/nussknacker/engine/flink/util/transformer/aggregate/AggregateHelper.java
@@ -27,6 +27,11 @@ public class AggregateHelper implements Serializable {
             new FixedExpressionValue("#AGG.last", "Last"),
             new FixedExpressionValue("#AGG.countWhen", "CountWhen"),
             new FixedExpressionValue("#AGG.average", "Average"),
+            new FixedExpressionValue("#AGG.stddevPop", "StddevPop"),
+            new FixedExpressionValue("#AGG.stddevSamp", "StddevSamp"),
+            new FixedExpressionValue("#AGG.varPop", "VarPop"),
+            new FixedExpressionValue("#AGG.varSamp", "VarSamp"),
+            new FixedExpressionValue("#AGG.median", "Median"),
             new FixedExpressionValue("#AGG.min", "Min"),
             new FixedExpressionValue("#AGG.max", "Max"),
             new FixedExpressionValue("#AGG.sum", "Sum"),
@@ -46,6 +51,11 @@ public class AggregateHelper implements Serializable {
     private static final Aggregator LAST = aggregates.LastAggregator$.MODULE$;
     private static final Aggregator COUNT_WHEN = aggregates.CountWhenAggregator$.MODULE$;
     private static final Aggregator AVERAGE = aggregates.AverageAggregator$.MODULE$;
+    private static final Aggregator STDDEV_POP = aggregates.PopulationStandardDeviationAggregator$.MODULE$;
+    private static final Aggregator STDDEV_SAMP = aggregates.SampleStandardDeviationAggregator$.MODULE$;
+    private static final Aggregator VAR_POP = aggregates.PopulationVarianceAggregator$.MODULE$;
+    private static final Aggregator VAR_SAMP = aggregates.SampleVarianceAggregator$.MODULE$;
+    private static final Aggregator MEDIAN = aggregates.MedianAggregator$.MODULE$;
     private static final Aggregator APPROX_CARDINALITY = HyperLogLogPlusAggregator$.MODULE$.apply(5, 10);
 
     // Instance methods below are for purpose of using in SpEL so your IDE can report that they are not used.
@@ -66,6 +76,13 @@ public class AggregateHelper implements Serializable {
 
     public Aggregator countWhen = COUNT_WHEN;
     public Aggregator average = AVERAGE;
+
+    public Aggregator stddevPop = STDDEV_POP;
+    public Aggregator stddevSamp = STDDEV_SAMP;
+    public Aggregator varPop = VAR_POP;
+    public Aggregator varSamp = VAR_SAMP;
+
+    public Aggregator median = MEDIAN;
 
     public Aggregator approxCardinality = APPROX_CARDINALITY;